@@ -3,27 +3,52 @@
 </template>
 
 <script setup>
-// import { useRoute } from "vue-router"
-import { onMounted } from "vue"
-// import { storeToRefs } from "pinia"
-// import { useStorePinia } from "@/store"
+    // import { onMounted } from 'vue'
+    // import { useRoute, useRouter } from 'vue-router'
+    // import { storeToRefs } from 'pinia'
+    // import { useStorePinia } from '@/store'
+    //
+    // const router = useRouter()
+    // const route = useRoute()
+    // const store = useStorePinia()
+    // const { count } = storeToRefs(store)
+    //
+    // const navigatorFun = (name) => {
+    //     router.push({
+    //         name,
+    //     })
+    // }
+    //
+    // onMounted(() => {
+    //     console.log(route, route.path)
+    // })
+    import { onMounted } from 'vue'
+    import { storeToRefs } from 'pinia'
+    import { useStorePinia } from '@/store'
 
-// const router = useRouter()
-// const route = useRoute()
-// const store = useStorePinia()
-// const { count } = storeToRefs(store)
+    const store = useStorePinia()
+    const { documentClientWidth, documentClientHeight, isExpand, isMobile, resizeNumber } = storeToRefs(store)
 
-// const navigatorFun = (name) => {
-//     router.push({
-//         name,
-//     })
-// }
+    onMounted(() => {
+        updateDocumentClientHeight()
+    })
 
-onMounted(() => {
-    // console.log(route, route.path)
-})
+    function updateDocumentClientHeight () {
+        documentClientWidth.value = document.documentElement['clientWidth']
+        documentClientHeight.value = document.documentElement['clientHeight']
+        isExpand.value = documentClientWidth.value <= 750 ? false : true
+        isMobile.value = documentClientWidth.value <= 750 ? true : false
+
+        window.addEventListener('resize', () => {
+            documentClientWidth.value = document.documentElement['clientWidth']
+            documentClientHeight.value = document.documentElement['clientHeight']
+            resizeNumber.value++
+            isExpand.value = documentClientWidth.value <= 750 ? false : true
+            isMobile.value = documentClientWidth.value <= 750 ? true : false
+        }, false)
+    }
 </script>
 
 <style lang="scss">
-@import '@/assets/style/index.scss'
+    @import '@/assets/style/index.scss'
 </style>
