@@ -6,7 +6,7 @@
 
         <div class="site-navbar__body fix">
             <div class="site-navbar__body-left">
-                <el-button :icon="isExpand? 'Fold' : 'Expand'" class="expand-btn" @click="isExpand = !isExpand"></el-button>
+                <el-button :icon="isExpand ? 'Fold' : 'Expand'" class="expand-btn" @click="isExpand = !isExpand"></el-button>
             </div>
 
             <el-menu class="site-navbar__menu" mode="horizontal" :ellipsis="false">
@@ -16,11 +16,17 @@
                 <el-menu-item class="site-navbar__avatar">
                     <el-dropdown :show-timeout="0" placement="bottom">
                         <span class="el-dropdown-link">
-                            <img src="~@/assets/img/avatar.png" :alt="userInfo.account">{{ userInfo.account }}
+                            <div class="img-box">
+                               <img src="@/assets/img/avatar.png" alt="" />
+                            </div>
+                            <span class="txt">{{ userInfo.account }}</span>
+                            <el-icon>
+                                <CaretBottom />
+                            </el-icon>
                         </span>
                         <template #dropdown>
                             <el-dropdown-menu >
-                                <el-dropdown-item @click="navigatorFun">用户信息</el-dropdown-item>
+                                <el-dropdown-item @click="resetPassword">修改密码</el-dropdown-item>
                                 <el-dropdown-item @click="logoutHandle">退出</el-dropdown-item>
                             </el-dropdown-menu>
                         </template>
@@ -33,17 +39,17 @@
 
 <script setup>
 import { onMounted, ref } from 'vue'
-import { ElMessageBox } from 'element-plus'
+import { ElLoading, ElMessageBox } from 'element-plus'
 import { clearLoginInfo, Fullscreen } from '@/utils'
-import { storeToRefs } from "pinia"
-import { useStorePinia } from "@/store"
+import { storeToRefs } from 'pinia'
+import { useStorePinia } from '@/store'
 import { useRouter } from 'vue-router'
-import { Fold, Expand } from '@element-plus/icons-vue'  // eslint-disable-line
+import { Fold, Expand, CaretBottom } from '@element-plus/icons-vue'  // eslint-disable-line
 
 const store = useStorePinia()
 const { userInfo, isExpand } = storeToRefs(store)
 const router = useRouter()
-const isFullscreen = ref(false);
+let isFullscreen = ref(false)
 
 onMounted(() => {
     document.addEventListener('fullscreenchange', () => {
@@ -52,25 +58,27 @@ onMounted(() => {
 })
 
 // 退出
-const logoutHandle = () => {
+function logoutHandle () {
     ElMessageBox.confirm(`确定进行[退出]操作?`, '提示', {
         confirmButtonText: '确定',
         cancelButtonText: '取消',
         type: 'warning'
     }).then(() => {
+        ElLoading.service({ lock: true })
         clearLoginInfo()
-        router.push({ name: 'login' })
         setTimeout(() => {
             location.reload()
-        }, 0)
+        }, 60)
     }).catch(() => {})
 }
-// 跳转到用户信息页
-const navigatorFun = () => {
-    router.push({ name: 'userInfor' })
+// 修改密码
+function resetPassword () {
+    router.push({
+        path: '/changePassword',
+    })
 }
 // 全屏切换
-const changeFullscreen = () => {
+function changeFullscreen () {
     isFullscreen.value ? Fullscreen.exitFullscreen() : Fullscreen.launchFullscreen(document.documentElement)
 }
 </script>
