@@ -5,50 +5,55 @@ export const useStorePinia = defineStore('main', {
     state () {
         return {
             token: sessionStorage.getItem('token') || '',
-            userInfo: {},
+            userInfo: sessionStorage.getItem('userInfo') ? JSON.parse(sessionStorage.getItem('userInfo')) : {},
             permission: [],  //权限
             menuList: [],  //右侧菜单
             dynamicMenuRoutes: [],  //动态路由
             mainTabs: [],  //tab列表
-            mainTabsActiveName: '',  //tab选中值
+            mainTabsActiveName: 'Home',  //tab选中值
+            documentClientWidth: 0,
             documentClientHeight: 0,
+            resizeNumber: 0,
             isExpand: true,
+            isMobile: false,
             dictType: localStorage.getItem('dictType') ? JSON.parse(localStorage.getItem('dictType')) : {}, // 数据字典
         };
     },
     getters : {
         //右侧菜单选中值
         menuActiveName () {
-            let str = ''
+            let str = this.mainTabsActiveName
             this.dynamicMenuRoutes.forEach((value) => {
-                if (value.name == this.mainTabsActiveName) {
+                if (value.name == this.mainTabsActiveName && this.mainTabsActiveName != '') {
                     str = value.meta.menuId + ''
                 }
             })
+
             return str
         },
     },
     actions: {
-        setDictType (val) {
-            this.dictType = val
-            localStorage.setItem('dictType', JSON.stringify(val))
-        },
         setToken (val) {
-            this.token = val
+            this.token = val;
             sessionStorage.setItem('token', val)
         },
+        setUserInfo (val) {
+            this.userInfo = val;
+            sessionStorage.setItem('userInfo', JSON.stringify(val))
+        },
         updateCommonStore (name, val) {
-            this[name] = val
+            this[name] = val;
         },
         resetStore () {
+            this.token = '';
             sessionStorage.removeItem('token')
-            this.token = ''
-            this.userInfo = {}
-            this.permission = []
-            this.menuList = []
-            this.dynamicMenuRoutes = []
-            this.mainTabs = []
-            this.mainTabsActiveName = ''
+            sessionStorage.removeItem('userInfo')
+            this.userInfo = {};
+            this.permission = [];
+            this.menuList = [];
+            this.dynamicMenuRoutes = [];
+            this.mainTabs = [];
+            this.mainTabsActiveName = '';
         },
         async getUserInfo () {
             const { data } = await userGetUserInfo();
@@ -57,7 +62,7 @@ export const useStorePinia = defineStore('main', {
             }
             this.permission = data.permission
 
-            return data           
+            return data
         },
     },
 });
