@@ -1,11 +1,13 @@
 <template>
-    <iframe
-        :src="route.meta.iframeUrl"
-        width="100%"
-        height="100%"
-        frameborder="0"
-        scrolling="yes">
-    </iframe>
+    <div class="table-flex-page">
+        <iframe
+            :src="route.meta.iframeUrl"
+            width="100%"
+            height="100%"
+            frameborder="0"
+            scrolling="yes">
+        </iframe>
+    </div>
 </template>
 
 <script setup>
