@@ -49,11 +49,10 @@
 import { onMounted, ref, nextTick } from 'vue'
 import indexAddOrUpdate from './index-add-or-update.vue'
 import { orderOrderPage, orderOrderDelete } from '@/api/order'
-import { ElLoading, ElMessage, ElMessageBox } from 'element-plus'
+import { ElLoading, ElMessage, ElMessageBox, dayjs } from 'element-plus'
 import { Plus } from '@element-plus/icons-vue'
 import { deepCopy } from '@/utils/index'
 import { commonMixin } from '@/mixins/common'
-const dayjs = require('dayjs')
 
 const { codeToLabelComputed } = commonMixin()
 const defaultDataForm = {
