@@ -1,47 +1,50 @@
 <template>
-    <!-- 角色管理 -->
-    <el-form :inline="true" :model="formData" @submit.prevent @keyup.enter="searchFun">
-        <el-form-item label="字典名称">
-            <el-input v-model="formData.name" placeholder="请输入" clearable class="inp-dom" />
-        </el-form-item>
-        <el-form-item label="字典类型">
-            <el-input v-model="formData.type" placeholder="请输入" clearable class="inp-dom" />
-        </el-form-item>
-        <el-form-item>
-            <el-button type="primary" @click="searchFun">查询</el-button>
-            <el-button @click="resetFun()">重置</el-button>
-            <el-button @click="lisDataFun()">字典数据查询</el-button>
-        </el-form-item>
-    </el-form>
+    <div class="table-flex-page">
+        <el-form :inline="true" :model="formData" label-width="70px" @submit.prevent style="margin-bottom: 12px">
+            <el-form-item label="字典名称">
+                <el-input v-model="formData.name" placeholder="请输入" clearable class="inp-dom" />
+            </el-form-item>
+            <el-form-item label="字典类型">
+                <el-input v-model="formData.type" placeholder="请输入" clearable class="inp-dom" />
+            </el-form-item>
+            <el-form-item>
+                <el-button type="primary" @click="searchFun">查询</el-button>
+                <el-button @click="resetFun()">重置</el-button>
+                <el-button @click="lisDataFun()">字典数据查询</el-button>
+            </el-form-item>
+        </el-form>
 
-    <div style="margin-bottom: 12px;">
-        <el-button type="primary" :icon="Plus" @click="addOrUpdateFun()">新增</el-button>
+        <div class="table-op-box" style="margin-bottom: 12px">
+            <div class="box-l">
+                <el-button type="primary" @click="addOrUpdateFun()"><el-icon><Plus /></el-icon><span>新增</span></el-button>
+            </div>
+        </div>
+
+        <el-table header-cell-class-name="table-cell-header" :data="dataList" border v-loading="dataListLoading" style="width: 100%;">
+            <el-table-column type="index" header-align="center" align="center" label="序号" width="70"></el-table-column>
+            <el-table-column prop="name" label="字典名称"></el-table-column>
+            <el-table-column prop="type" label="字典类型"></el-table-column>
+            <el-table-column prop="status" label="状态">
+                <template #default="scope">
+                    <span v-if="scope.row.status == 1">开启</span>
+                    <span v-else>关闭</span>
+                </template>
+            </el-table-column>
+            <el-table-column prop="createTime" label="创建时间"></el-table-column>
+            <el-table-column label="操作" fixed="right">
+                <template #default="scope">
+                    <el-button type="primary" link @click="addOrUpdateFun(scope.row.id)">编辑</el-button>
+                    <el-button type="primary" link @click="lisDataFun(scope.row.id)">字典数据</el-button>
+                    <el-button type="danger" link @click="delFun(scope.row)">删除</el-button>
+                </template>
+            </el-table-column>
+        </el-table>
+
+        <el-pagination class="el-pagination" @size-change="sizeChangeHandle" @current-change="currentChangeHandle"
+                       :current-page="formData.pageIndex" :page-sizes="[10, 20, 50, 100]" :page-size="formData.pageSize"
+                       :total="formData.totalPage" layout="total, sizes, prev, pager, next, jumper">
+        </el-pagination>
     </div>
-
-    <el-table header-cell-class-name="table-cell-header" :data="dataList" border v-loading="dataListLoading" style="width: 100%;">
-        <el-table-column type="index" header-align="center" align="center" label="序号" width="70"></el-table-column>
-        <el-table-column prop="name" label="字典名称"></el-table-column>
-        <el-table-column prop="type" label="字典类型"></el-table-column>
-        <el-table-column prop="status" label="状态">
-            <template #default="scope">
-                <span v-if="scope.row.status == 1">开启</span>
-                <span v-else>关闭</span>
-            </template>
-        </el-table-column>
-        <el-table-column prop="createTime" label="创建时间"></el-table-column>
-        <el-table-column label="操作" fixed="right">
-            <template #default="scope">
-                <el-button type="primary" link @click="addOrUpdateFun(scope.row.id)">编辑</el-button>
-                <el-button type="primary" link @click="lisDataFun(scope.row.id)">字典数据</el-button>
-                <el-button type="primary" link @click="delFun(scope.row)">删除</el-button>
-            </template>
-        </el-table-column>
-    </el-table>
-
-    <el-pagination class="el-pagination" @size-change="sizeChangeHandle" @current-change="currentChangeHandle"
-                   :current-page="formData.pageIndex" :page-sizes="[10, 20, 50, 100]" :page-size="formData.pageSize"
-                   :total="formData.totalPage" layout="total, sizes, prev, pager, next, jumper">
-    </el-pagination>
 
     <!-- 新增和编辑弹框 -->
     <indexAddOrUpdate ref="indexAddOrUpdateRef" @refreshDataList="searchFun" @close="indexAddOrUpdateVisible= false"
@@ -53,14 +56,14 @@
 </template>
 
 <script setup>
-import { onMounted, ref, nextTick } from 'vue'
+import { onMounted, ref, nextTick, defineOptions } from 'vue'
 import { dictTypePage, dictTypeDelete } from '@/api/system'
 import indexAddOrUpdate from './index-add-or-update.vue'
 import indexData from './index-data.vue'
 import { dayjs, ElLoading, ElMessage, ElMessageBox } from 'element-plus'
-import { Plus } from '@element-plus/icons-vue'
 import { deepCopy } from '@/utils/index'
 
+defineOptions({ name: 'SystemDictIndex' })
 const defaultFormData = {
     pageIndex: 1,
     pageSize: 10,
